model/requirements.txt
# Requirements for GroundingDINO auto-labeling
transformers
torch
Pillow
tqdm
numpy
